Fixed integrationtests.
Added integration tests back to pipeline.

Fixed an error in integration test which which resulted in a logged out user for cypress still being logged in.
This happens because the API requests to the usermgt api has stored session cookies and cypress assumed that the user is still logged in.

Update test-libary to v3.0.2 in which the bug with the warpemnu checkbox
is fixed.