#115 Extract startup functions for later test; add Cloudogu startup logo
The landscape of functionality of `startup.sh` is pretty much untested,
mostly because it is not modular. This commit re-uses the already
existing structure and moves the functionality into functions without
change of the business logic.

Source guards take care that later tests can properly test single
functions without running the whole dogu.