#115 Simplify Docker build stage; remove root user; expand .dockerignore
The Docker build process previously consisted of three build stages:
- BUILDER: usermgt binary building
- BUILDER: tomcat downloading
- PROD: dogu building

All three stages did not worked up to their potential as lots of the
statements were glued together in repeated ways, `cd` was used where
WORKDIR would provide better filesystem insights and so on.

This commit introduces a new Docker builder stage which combines the two
builder stages and prepares the output filesystem which will be simply
copied into the production stage. Because the builder stages do not care
about layer sizes the number of RUN commands is not as limited as in a
production stage (where the number should be reduced as possible)

Command used to determine complexity and size of the images:
```
docker run --rm -it     -v /var/run/docker.sock:/var/run/docker.sock     wagoodman/dive:latest registry.cloudogu.com/official/usermgt:1.12.1-1
```
The total image size was reduced by 29 MB (326 MB original, 297 MB
rebuild).

Because of the way the JRE truststore was created, in the past a root
user was necessary. This commit adds also a higher level of security
because the container does no longer basically runs as root because the
truststore is now moves from /etc/.. (root owned) to /opt/.. (tomcat
owned)