#117 Fix file permission bug of /var/lib/usermgt resources
In #115/#116 was a bug introduced that occurred when a previous dogu
version was installed. Prior to this commit, this behaviour can be
reproduced like this:

```bash
cesapp install official/usermgt 1.12.1-1
cesapp start usermgt
cesapp build .
```

The problem lies in the change of the container user of the previous
dogu version, that is: root. In a previous version, `startup.sh` will
have created a directory `/var/lib/usermgt/conf/` which should contain
runtime resources for the Java application. Anyhow, by that time,
`startup.sh` was running with root permissions, thus creating the above
directory with root permissions only.

This problem can be countered with three solutions:

1. remove the directory outside the container
2. change ownership of the directory during an upgrade script
3. create a directory alternative to `/var/lib/usermgt/conf/`

Solution 1 to clean up outside the ownership with in the dogu volume
does not scale well because before/during/after the dogu upgrade the
administrator needs to do this or create a script (when this is actually
the dogu's task)

Solution 2 is in the realm of the possibility when the script can be run
with root permissions. Thus, this action must be executed before this
version because this version will only run with user 1000:1000. Running
as pre-upgrade script is possible though because all previous versions
should run with the root user instead.

Solution 3 creates an alternative app configuration directory where the
dogu is guaranteed to maintain permissions of UID/GID 1000:1000. Without
solution 2 ("get somehow rid of the root owned directory") this solution
does work while leaving ~3k bytes of files. This is not exactly elegant
but bearable. Solution 2 may be implemented in later commits.

This commit re-uses the already existing environment variable
`UNIVERSEADM_HOME` and replaces hardcoded occurrences with the env var
or with a respective Java property (set via `setenv.sh`)

Since `mkdir -p` already creates a directory conditionally to whether it
exists already or not, this commit also removes the surrounding bash
condition.