#153 Fix syntax in Then-step definition, add comment on weird behavior
IntelliJ showed some syntax errors in "the user {string} was created"
definitions where argument types and some parenthesis were missing.
This comment fixes this by adding these things.

This commit also adds a comment on the When code for "the user opens the
user import summary details page". The function defines a parameter but
the descriptions does not allow for a parameter to be given in the first
place. This all is made even weirder as the When then leads to an
HTTP401 but the following Then ask for an access denied message. AND YET
THE TEST PASSES! I have no actual solution beyond fully rewriting these
steps as following:

1. login as manager and create import
2. remember UUID of the import URI
3. logout and login as nomanager
4. try to achieve access to the import URI
5. expect an access denied message.

This commit does not fix this though because it is way out of scope.

Co-authored-by: Laura Gorzitze <laura.gorzitze@cloudogu.com>