#157 move code to first usage, replace exception, introduce Assert4J
In order to add to the test coverage some things were screwy in the
existing codebase. `getInputStream()` threw an exception from a SAML
package which did not make any sense. This commit replaces this with
Java's standard `IllegalArgumentException`. I am not a huge friend of
this replacement, mainly because there are better substitutions, but it
is closest to the original code.

This commit also adds the Assert4J framework to the testing scope
because it adds A F%/&!ING LOT of ease-of-use for assertions. This also
deprecates some homebrew assertion methods in ImportHandlerTest.java