Revert "#157 Update ch.qos.logback in POM.xml to remove CVEs"
This reverts commit dae24627f9020770e36ee9f8bbc4abc33e21cdfb.

Okay, I need to vent here for a bit. This commit must be rolled back
because the previous commit tried to smoothly get rid of a CVE with a
score of 9.8 (and ofc some others as the used libraries are dead-old).

As it turns out this does not work. It will build but I find no way to
properly XML-configure logback along with SLF4J. Instead nothing will be
logged and during start-up the dogu logs something like "SLF4J will fall
back to a no-op logger" or something similar. In the end, the usermgt
works fine but absolutely no (like NONE) logs will be printed to the
container output stream.

Updating all versions including logback and SLF4J does not help, also
not updating the XML config to the most current recommendation. Instead
either no logs or even build failures where NoClassDefFoundError is a
thing for the LoggerFactory implementation.

Using files directly from Maven central ALSO does not help because some
old-ass com.github.legman library will be used that is only available in
the scm-manager specific repository. It is completely unknown to maven
central because it is sth like a Cloudogu fork?

Either way, the underlying libraries MUST be replaced at some point to
deliver the business functionality in order to resolve CVEs within the
logging frameworks (and we all KNOW this was a big deal in the not-so-
far history).

Anyhow, if some lone developer finds this commit message: Godspeed, I
have you in my heart but you really have to deal with the original
business requirements to solve this spaghetti mess, or rewrite
everything from scratch. The usermgt isn't that special. Blessings and
peace out!