#157 Hide HTTP 409 message from user and show generic error instead
The original HTTP 409 response from the usermgt backend cannot be easily
exchanged because the responding class is a generic one. Changing this
would disturb other functionalities.

The one thing that can be done is to hide the none-sense message and
replace it with a generic message. Real Haxx0rz will still be able to
find out other users' email addresses from within the instance if they
aren't allowed to.