#157 Remove incomplete test
This test did not work out, and it was a mistake to commit it from the
beginning. I could not successfully test the import handler routing
which takes care of different exception states without expanding the
methods visibility into the default scope because this might come with
security issues.

Currently I regard this code untestable as of now, please prove me wrong
and make it work but I have no idea how. What I currently tried:
- different flavors of having the mock manager throw an exception during
  the `create()` function