#157 Mark input fields with invalid data after backend validation
This commit properly refines validation constraint from HTTP response
and harmonizes the way how HTTP requests and responses for users and
accounts are done.

According to the Axios docs, axios treats every response as an error if
it falls out of the HTTP 2xx range, so the replacement in the account
service looks reasonable.