#158 Change "assign member" labels for better usability
This change repeats a similar change from commit 3a3c80c9 which only
targeted a user's group assignment. This commit changes the same thing
from the other way around for the same reason: The wording might give
the impression to create a user or a group. This impression is clearly
wrong because it takes more than just a name to both a user/group and to
assign them properly.