#202 Return other statuses than HTTP 401 for connection/request errors
This commit reduces the probability of HTTP 401 Unauthorized status
codes which previously happened also when almost ANY errors occurred,
even those technical ones outside the realm of authentication.

This commit also added some tests which made some changes of class
visibility necessary. A new constructor allows for testability while the
default ie. the public constructor remains unchanged. A factory must be
created as new HTTP connections are created both for TGT as well as for
ST requests. This is hard to test so the factory is unavoidable.

For a better mock experience of final classes, Mockito library versions
must be bumped up.